Upgrade to Docker Compose v2 spec
Future-compatible with v2.1 and v3 right now.

Using the depends_on feautre still does not replace the need for the sleep on startup of taiga web container. Although, it foces the postgres to startup first, so I moved it to the bottom of the docker-compose.yml file.

This ugprade removed the exposure of the shared envirionment variables, so removed the remaining use of POSTGRES_ENV_POSTGRES_PASSWORD. You must now always specify the password.

Added the pgdata volume to example docker-compose.yml postgres container definition.